Separate contents of Systematic and Logging
